refactor: extract LocalPrivacyDataManager.swift from PrivacyDataExporter.swift
Extracts the `LocalPrivacyDataManager` struct + its
`LocalPrivacyDataManaging` conformance extension out of
PrivacyDataExporter.swift into its own file. PrivacyDataExporter.swift
retains only the exporter struct + its conformance extension.

Byte-identical move of lines 170-209 (includes the load-bearing
Sendable rationale doc-comment). Only imports Foundation.

PrivacyDataExporter.swift: 209 → 168 lines. Tests: 667.
